Move scheduler configuration into the scheduler subproject (#3585)
